refactor: :recycle: move icon component under src to be common
